Move callout to common folder and adjust default callout message
